refactor: introduce HandlerResult and split inline handlers
HandlerResult is a value object that lets handlers express three distinct
outcomes — append a single node, append a list of nodes (concat), or skip
the element entirely — without nil-punning or special return values.
`HandlerResult.none` replaces the previous nil sentinel; `append_to`
encapsulates the concat-vs-push decision that was previously scattered.

Split the inline handler into two focused modules under
`handlers/inline/`:
- RichHtmlRenderer: produces the HTML-flavored inline output (was inline.rb)
- TextExtractor: produces plain-text inline output for attribute values

The inline.rb entrypoint now dispatches to both modules via the registry,
keeping the handler_registry free of inline-specific knowledge.
